requirements.txt
djangorestframework
django
python-decouple
gunicorn
psycopg2-binary
django-cors-headers
sentry-sdk
django_model_utils
pillow
django_filter